Add H2 database dependency for testing, update configuration properties, and enhance audit entity with additional getters
- Added H2 database dependency in pom.xml for testing purposes.
- Updated TiersPayantStockageProperties configuration prefix to use hyphens.
- Enhanced EvenementAuditJpaEntity by adding getter methods for various fields.
- Modified EvenementAuditJpaRepository to include a new search method with filtering capabilities.
- Updated AjouterCodeBarresCommand to include organisationId and adjusted related use cases and controllers.
- Refactored CategorieController and ProduitController to accommodate new data structures and validation rules.
- Improved patient management by changing assurance coverage type from Double to BigDecimal for better precision.
- Added new methods in various repositories to check existence by organisationId and ID.
- Enhanced SequenceNumerotationJpaEntity with reset logic for sequence counters based on frequency.
- Removed unnecessary component annotation from StocksReceptionProperties class.

+++ backend/src/main/java/cm/pharma/contexts/ordonnances_dossier_patient/application/command/CreerPatientUseCase.java
@@ -7,6 +7,7 @@ import cm.pharma.contexts.ordonnances_dossier_patient.infrastructure.persistence
 import cm.pharma.contexts.ordonnances_dossier_patient.infrastructure.persistence.jpa.PatientMedicalJpaEntity;
 import cm.pharma.contexts.ordonnances_dossier_patient.infrastructure.persistence.jpa.PatientMedicalJpaRepository;
 import cm.pharma.shared.domain.BusinessRuleViolationException;
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.Map;
 import java.util.Objects;
@@ -75,7 +76,7 @@ public class CreerPatientUseCase {
             String adresse,
             String assuranceOrganismeNom,
             String assuranceNumeroAdherent,
-            Double assuranceTauxCouverture,
+            BigDecimal assuranceTauxCouverture,
             UUID creePar,
             String posteNom
     ) {
